refactor: changed name of module ann to lhs
