+++ src/search.rs
@@ -10,7 +10,7 @@ pub use methods::message_scan::MessageScan;
 pub use methods::similar_diff::ANNMatch;
 pub use methods::similar_diff::HNSWSearch;
 pub use methods::similar_diff::SimilarityDiffMatch;
-pub use methods::ann::traditional_lsh::TraditionalLSH;
+pub use methods::lsh::traditional_lsh::TraditionalLSH;
 
 #[derive(Debug, Hash, PartialEq, Eq)]
 pub struct CherryAndTarget {
