+++ src/search/methods.rs
@@ -1,4 +1,4 @@
-pub mod ann;
+pub mod lsh;
 pub mod exact_diff;
 pub mod message_scan;
 pub mod similar_diff;
