+++ src/search/methods/lsh.rs
