+++ src/search/methods/lsh/preprocessing.rs
@@ -291,7 +291,7 @@ impl MinHash {
 #[cfg(test)]
 mod tests {
     use crate::git::IdeaPatch;
-    use crate::search::methods::ann::preprocessing::{
+    use crate::search::methods::lsh::preprocessing::{
         preprocess_texts, shingle_diff, MinHash, ShingledText, Signature, Vocabulary,
     };
     use crate::Diff;
